Go: miscellaneous documentation improvements
Improves docstrings for the go public functions in some areas. Also
reorganizes the order of types in mcap.go to match the order in the
spec, removes some unused types, and splits the custom reader/writer
types into their own files.
